Noetic update
This update allows the ARIAC 2019 to work in ROS Noetic on Ubuntu 20.04.  There update is not complete.  The conversion of the configuration scripts failed because of a problem with the empy package for python that seems to fail internally.

To manage this problem, the dynamically generated files for the project that are created by gear.py have been included statically.  The most important implication is that the configuration of the environment and the orders cannot be updated.