Clearer fixes for the empy bug
There is a patch for em.py that can be applied to make the original ARIAC 2019 package work.  It is a one line fix.  If it is applied, then the ARIAC 2019 will work as normal.  Otherwise, the launch file in this package can take an argument to use static files included as a work around instead.

The next version of empy due to be released will include the fix to the bug, but it is unclear when it will be released.  Once it is, though, the workarounds will be removed.